More duration stuff
Also new flow:update command to pull latest assets.

+++ database/migrations/2022_02_26_000000_add_duration_to_flow_events_table.php
@@ -9,7 +9,7 @@ class AddDurationToFlowEventsTable extends Migration
     public function up()
     {
         Schema::table('flow_events', function (Blueprint $table) {
-            $table->integer('duration_ms')->default(0);
+            $table->float('duration_ms')->default(0.0);
         });
     }
